fix(homepage): make homepage configurable with NEXT_PUBLIC_SHOW_COMING_SOON env var
- Extract coming soon page to reusable component
- Update root page.tsx to conditionally render coming soon or full homepage
- Default to showing actual landing page (home-3d) in production
- Set NEXT_PUBLIC_SHOW_COMING_SOON=true to revert to coming soon mode
- Add appropriate metadata for both modes

Fixes MEDIUM issue #6 (homepage in coming soon mode while rest of site is public)
